laboratorio_2/requirements.txt
mysql-connector-python==9.0.0
python-decouple==3.8
